+++ test/test_init.py
@@ -1,6 +1,6 @@
 import unittest
 from datetime import datetime, timedelta, timezone
-from init.handler import get_file_ids_to_be_processed
+from init.helpers import get_file_ids_to_be_processed
 
 test_config = {
   "process-bucket-csv-basepath": "csv",
